removed allowed http methods from handlers
