+++ TKeazirian.HTTPServer/Handler/Handler.cs
@@ -5,11 +5,6 @@ using Request;
 
 public abstract class Handler
 {
-    public virtual List<string> AllowedHttpMethods()
-    {
-        throw new Exception("Method not implemented");
-    }
-
     public virtual Response HandleResponse(Request request)
     {
         throw new Exception("Method not implemented");
