+++ TKeazirian.HTTPServer/Handler/ResourceNotFoundHandler.cs
@@ -5,11 +5,6 @@ using Request;
 
 public class ResourceNotFoundHandler : Handler
 {
-    public override List<string> AllowedHttpMethods()
-    {
-        return new List<string>() { "GET" };
-    }
-
     public override Response HandleResponse(Request request)
     {
         string body = "The resource cannot be found";
